updated api endpoints and endpoints

+++ next_app/app/api/admin/products/[id]/route.ts
@@ -1,9 +1,35 @@
 import { NextResponse } from 'next/server'
 import { PrismaClient } from '@prisma/client'
+import { getServerSession } from 'next-auth'
+import { authOptions } from '../../../auth/[...nextauth]/route'
 
 const prisma = new PrismaClient()
 
+export async function PUT(request: Request, { params }: { params: { id: string } }) {
+  const session = await getServerSession(authOptions)
+  if (!session?.user?.is_admin) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 403 })
+  }
+
+  try {
+    const { name, description, price, image } = await request.json()
+    const updatedProduct = await prisma.product.update({
+      where: { id: params.id },
+      data: { name, description, price, image },
+    })
+    return NextResponse.json(updatedProduct)
+  } catch (error) {
+    console.error('Error updating product:', error)
+    return NextResponse.json({ error: 'Error updating product' }, { status: 500 })
+  }
+}
+
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+  const session = await getServerSession(authOptions)
+  if (!session?.user?.is_admin) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 403 })
+  }
+
   try {
     await prisma.product.delete({
       where: { id: params.id },
